Structure finished, some codes about CLI and Config files, tests of RakNet implementation on server (not in git tracking files)
