+++ server/Headers/InputControllers/CControllerState.h
@@ -1,6 +1,8 @@
-#if !defined(_COOPSERVER_INPUTCONTROLLERS_CONTROLLERSTATE_H_) || !defined(_COOPSERVER_CONTROLLERSTATE_H_)
+#if !defined(_COOPSERVER_INPUTCONTROLLERS_CONTROLLERSTATE_H_) || !defined(_COOPSERVER_CCONTROLLERSTATE_H_)
 #define _COOPSERVER_INPUTCONTROLLERS_CONTROLLERSTATE_H_
-#define _COOPSERVER_CONTROLLERSTATE_H_
+#define _COOPSERVER_CCONTROLLERSTATE_H_
+
+#include <cstdint>
 
 //#pragma pack(1)
 struct CCompressedControllerState
