+++ server/Headers/Network/CPacketListener.h
@@ -1,6 +1,6 @@
-#if !defined(_COOPSERVER_NETWORK_PACKETLISTENER_H_) || !defined(_COOPSERVER_PACKETLISTENER_H_)
+#if !defined(_COOPSERVER_NETWORK_PACKETLISTENER_H_) || !defined(_COOPSERVER_CPACKETLISTENER_H_)
 #define _COOPSERVER_NETWORK_PACKETLISTENER_H_
-#define _COOPSERVER_PACKETLISTENER_H_
+#define _COOPSERVER_CPACKETLISTENER_H_
 
 #include <vector>
 #include <algorithm>
