+++ server/Headers/Player/CPlayer.h
@@ -1,6 +1,6 @@
-#if !defined(_COOPSERVER_PLAYER_Player_H_) || !defined(_COOPSERVER_CPLAYER_H_)
+#if !defined(_COOPSERVER_PLAYER_PLAYER_H_) || !defined(_COOPSERVER_CPLAYER_H_)
 #define _COOPSERVER_CPLAYER_H_
-#define _COOPSERVER_PLAYER_Player_H_
+#define _COOPSERVER_PLAYER_PLAYER_H_
 
 #include <string>
 #include "../../include/enet/enet.h"
@@ -12,20 +12,18 @@
 class CPlayer
 {
 	public:
-		CPlayer(ENetPeer* peer, int playerid);
-
-		ENetPeer* m_pPeer;
 		int m_iPlayerId;
-		char m_Name[32 + 1] = { 0 };
-		bool m_bIsHost;
-		unsigned char m_nSeatId;
 		int m_nVehicleId;
-		bool m_bCorrectVersion = false;
 		float m_afStats[14];
-		unsigned int m_anModelKeys[10];
-		unsigned int m_anTextureKeys[18];
 		float m_fFatStat;
 		float m_fMuscleStat;
+		unsigned int m_anModelKeys[10];
+		unsigned int m_anTextureKeys[18];
+		bool m_bCorrectVersion = false;
+		bool m_bIsHost;
+		char m_Name[32 + 1] = { 0 };
+		unsigned char m_nSeatId;
+		ENetPeer* m_pPeer;
 		CVector m_vecWaypointPos{};
 		
 		struct {
@@ -33,6 +31,8 @@ class CPlayer
 			uint8_t bClothesModified : 1;
 			uint8_t bWaypointModified : 1;
 		} m_ucSyncFlags;
+
+		CPlayer(ENetPeer* peer, int playerid);
 
 		std::string GetName();
 		void RemoveFromVehicle();
