+++ server/Headers/Player/CPlayerPackets.h
@@ -1,4 +1,12 @@
+#if !defined(_COOPSERVER_PLAYER_CPLAYERPACKETS_H_) || !defined(_COOPSERVER_CPLAYERPACKETS_H_)
+#define _COOPSERVER_PLAYER_CPLAYERPACKETS_H_
+#define _COOPSERVER_CPLAYERPACKETS_H_
 
+#include "CPlayer.h"
+#include "CPlayerManager.h"
+#include "../Vector/CVector.h"
+//#include "../InputControllers/CControllerState.h"
+#include "../Network/NetworkEntityType.h"
 
 class CPlayerPackets
 {
@@ -567,3 +575,5 @@ public:
 		}
 	};
 };
+
+#endif
