+++ server/Headers/Server/CLI/Commands/CServerCommandHandler.h
@@ -1,5 +1,6 @@
-#if !defined(_COOP_SERVER_CSERVERCOMMANDMANAGER_H_)
-#define _COOP_SERVER_CSERVERCOMMANDMANAGER_H_
+#if !defined(_COOPSERVER_SERVER_CLI_COMMANDS_SCOMMANDMANAGER_H_) || !defined(_COOPSERVER_SERVERCOMMANDMANAGER_H_)
+#define _COOPSERVER_SERVER_CLI_COMMANDS_SCOMMANDMANAGER_H_
+#define _COOPSERVER_SERVERCOMMANDMANAGER_H_
 
 #include <iostream>
 #include <cstdio>
