+++ server/Headers/Vehicle/CVehicle.h
@@ -1,31 +1,31 @@
-#pragma once
+#if !defined(_COOPSERVER_VEHICLE_VEHICLE_H_) || !defined(_COOPSERVER_CVEHICLE_H_)
+#define _COOPSERVER_VEHICLE_VEHICLE_H_
+#define _COOPSERVER_CVEHICLE_H_
 
-#ifndef _CVEHICLE_H_
-	#define _CVEHICLE_H_
 #include <iostream>
 #include <vector>
 #include <algorithm>
 
-#include "CVector.h"
-#include "CPlayerManager.h"
+#include "../Vector/CVector.h"
+#include "../Player/CPlayerManager.h"
 
 class CVehicle
 {
 	public:
-		CVehicle(int vehicleid, unsigned short model, CVector pos, float rot);
 		
 		int m_nVehicleId;
-		CPlayer* m_pSyncer = nullptr;
 		unsigned short m_nModelId;
+		unsigned char m_nPrimaryColor;
+		unsigned char m_nSecondaryColor;
+		unsigned char m_damageManager_padding[23] = { 0 };
+		CPlayer* m_pSyncer = nullptr;
 		CVector m_vecPosition;
 		CVector m_vecRotation;
 		CVector m_vecVelocity;
 		CPlayer* m_pPlayers[8] = { nullptr };
-		unsigned char m_nPrimaryColor;
-		unsigned char m_nSecondaryColor;
-		unsigned char m_damageManager_padding[23] = { 0 };
 		std::vector<int> m_pComponents;
-	
+
+		CVehicle(int vehicleid, unsigned short model, CVector pos, float rot);
 		void ReassignSyncer(CPlayer* newSyncer);
 		void SetOccupant(uint8_t seatid, CPlayer* player);
