+++ server/Headers/Vehicle/VehicleDoorState.h
@@ -1,7 +1,6 @@
-#pragma once
-
-#ifndef _VEHICLEDOORSTATE_H_
-	#define _VEHICLEDOORSTATE_H_
+#if !defined(_COOPSERVER_VEHICLE_VEHICLEDOORSTATE_H_) || !defined(_COOPSERVER_CVEHICLEDOORSTATE_H_)
+#define _COOPSERVER_CVEHICLEDOORSTATE_H_
+#define _COOPSERVER_VEHICLE_VEHICLEDOORSTATE_H_
 
 struct VehicleDoorState
 {
